1.社团详情 2.推送打开Window 3.修复任务BUG
git-svn-id: https://39.108.144.9:8443/svn/xhandroid@82 5643548d-eeac-984e-88ce-79bbf109bbfc

+++ branches/2.0.0-final/Code/app/src/main/java/cn/flyexp/entity/MyTaskResponse.java
@@ -37,7 +37,7 @@ public class MyTaskResponse extends BaseResponse {
         @SerializedName("alt_phone")
         private String alt_phone;
         @SerializedName("fee")
-        private float fee;
+        private int fee;
         @SerializedName("phone")
         private String phone;
         @SerializedName("created_at")
@@ -179,11 +179,11 @@ public class MyTaskResponse extends BaseResponse {
             this.alt_phone = alt_phone;
         }
 
-        public float getFee() {
+        public int getFee() {
             return fee;
         }
 
-        public void setFee(float fee) {
+        public void setFee(int fee) {
             this.fee = fee;
         }
